feat(cmd/fetch): add Reddit fetch command
