+++ cmd/root.go
@@ -29,7 +29,7 @@ import (
 	"github.com/spf13/viper"
 )
 
-const version = "0.8.3"
+const version = "0.9.0"
 
 type rootOptions struct {
 	configFile  flag.FilePath
